extract Dirty into a shared mixin
