+++ lib/json_api_client/helpers.rb
@@ -1,20 +1,7 @@
 module JsonApiClient
   module Helpers
-    autoload :Associable, 'json_api_client/helpers/associable'
-    autoload :Attributable, 'json_api_client/helpers/attributable'
-    autoload :CustomEndpoints, 'json_api_client/helpers/custom_endpoints'
-    autoload :CustomHeaders, 'json_api_client/helpers/custom_headers'
-    autoload :DynamicAttributes, 'json_api_client/helpers/dynamic_attributes'
-    autoload :Initializable, 'json_api_client/helpers/initializable'
-    autoload :Inspectable, 'json_api_client/helpers/inspectable'
-    autoload :Linkable, 'json_api_client/helpers/linkable'
-    autoload :Relatable, 'json_api_client/helpers/relatable'
-    autoload :Paginatable, 'json_api_client/helpers/paginatable'
-    autoload :Parsable, 'json_api_client/helpers/parsable'
-    autoload :Queryable, 'json_api_client/helpers/queryable'
-    autoload :Requestable, 'json_api_client/helpers/requestable'
-    autoload :Schemable, 'json_api_client/helpers/schemable'
-    autoload :Serializable, 'json_api_client/helpers/serializable'
     autoload :Callbacks, 'json_api_client/helpers/callbacks'
+    autoload :Dirty, 'json_api_client/helpers/dirty'
+    autoload :DynamicAttributes, 'json_api_client/helpers/dynamic_attributes'
   end
 end
