+++ lib/json_api_client/resource.rb
@@ -10,6 +10,7 @@ module JsonApiClient
     include ActiveModel::Conversion
 
     include Helpers::DynamicAttributes
+    include Helpers::Dirty
 
     attr_accessor :last_result_set, :links, :relationships
     class_attribute :site,
